Структуризация кода
