+++ palettes/default.json
@@ -1,0 +1,1 @@
+["FF597D27", "FF6D9930", "FF7FB238", "FFAEA473", "FFD5C98C", "FFF7E9A3", "FF8C8C8C", "FFABABAB", "FFC7C7C7", "FFB40000", "FFDC0000", "FFFF0000", "FF7070B4", "FF8A8ADC", "FFA0A0FF", "FF757575", "FF909090", "FFA7A7A7", "FF005700", "FF006A00", "FF007C00", "FF737681", "FF8D909E", "FFA4A8B8", "FF6A4C36", "FF825E42", "FF976D4D", "FF4F4F4F", "FF606060", "FF707070", "FF645432", "FF7B663E", "FF8F7748", "FFB4B1AC", "FFDCD9D3", "FFFFFCF5", "FFB4B4B4", "FFDCDCDC", "FFFFFFFF", "FF985924", "FFBA6D2C", "FFD87F33", "FF7D3598", "FF9941BA", "FFB24CD8", "FF486C98", "FF5884BA", "FF6699D8", "FFA1A124", "FFC5C52C", "FFE5E533", "FF599011", "FF6DB015", "FF7FCC19", "FFAA5974", "FFD06D8E", "FFF27FA5", "FF353535", "FF414141", "FF4C4C4C", "FF6C6C6C", "FF848484", "FF999999", "FF35596C", "FF416D84", "FF4C7F99", "FF592C7D", "FF6D3699", "FF7F3FB2", "FF24357D", "FF2C4199", "FF334CB2", "FF483524", "FF58412C", "FF664C33", "FF485924", "FF586D2C", "FF667F33", "FF6C2424", "FF842C2C", "FF993333", "FF111111", "FF151515", "FF191919", "FFB0A836", "FFD7CD42", "FFFAEE4D", "FF409A96", "FF4FBCB7", "FF5CDBD5", "FF345AB4", "FF3F6EDC", "FF4A80FF", "FF009928", "FF00BB32", "FF00D93A", "FF5B3C22", "FF6F4A2A", "FF815631", "FF4F0100", "FF600100", "FF700200", "FF937C71", "FFB4988A", "FFD1B1A1", "FF703919", "FF89461F", "FF9F5224", "FF693D4C", "FF804B5D", "FF95576C", "FF4F4C61", "FF605D77", "FF706C8A", "FF835D19", "FFA0721F", "FFBA8524", "FF485225", "FF58642D", "FF677535", "FF703637", "FF8A4243", "FFA04D4E", "FF281C18", "FF31231E", "FF392923", "FF5F4B45", "FF745C54", "FF876B62", "FF3D4040", "FF4B4F4F", "FF575C5C", "FF56333E", "FF693E4B", "FF7A4958", "FF352B40", "FF41354F", "FF4C3E5C", "FF352318", "FF412B1E", "FF4C3223", "FF35391D", "FF414624", "FF4C522A", "FF642A20", "FF7A3327", "FF8E3C2E", "FF1A0F0B", "FF1F120D", "FF251610", "FF852122", "FFA3292A", "FFBD3031", "FF682C44", "FF7F3653", "FF943F61", "FF401114", "FF4F1519", "FF5C191D", "FF0F585E", "FF126C73", "FF167E86", "FF286462", "FF327A78", "FF3A8E8C", "FF3C1F2B", "FF4A2535", "FF562C3E", "FF0E7F5D", "FF119B72", "FF14B485", "FF464646", "FF565656", "FF646464", "FF597569", "FF6D9081", "FF7FA796"]
